+++ concourse-server/src/main/java/com/cinchapi/concourse/util/TMaps.java
@@ -27,7 +27,7 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 
 import com.cinchapi.concourse.server.plugin.data.Insertable;
-import com.cinchapi.concourse.server.plugin.data.LazyTrackingResultDataset;
+import com.cinchapi.concourse.server.plugin.data.LazyTrackingTObjectResultDataset;
 import com.google.common.base.Supplier;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -161,7 +161,7 @@ public final class TMaps {
     public static <E, A, V> void putResultDatasetOptimized(
             Map<E, Map<A, Set<V>>> map, E entity, Map<A, Set<V>> data) {
         if(map instanceof Insertable
-                && !(map instanceof LazyTrackingResultDataset)) {
+                && map.getClass() != LazyTrackingTObjectResultDataset.class) {
             Insertable<E, A, V> dataset = (Insertable<E, A, V>) map;
             for (Entry<A, Set<V>> entry : data.entrySet()) {
                 A attribute = entry.getKey();
